user detailed view

+++ Icesi-viajes/src/main/java/co/edu/icesi/viajes/domain/Rol.java
@@ -9,7 +9,7 @@ public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "nombre", nullable = false)
     private String nombre;
@@ -21,18 +21,18 @@ public class Rol {
     	
     }
     
-	public Rol(Long id, String nombre) {
+	public Rol(Integer id, String nombre) {
 		super();
 		this.id = id;
 		this.nombre = nombre;
 	}
 
 	// Getters and setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
